requirements_setup.txt
# Requirements required by this package:

brotlipy~=0.6.0
Django~=1.10.2
